fix: order-independent stable-id opacity; split oversized modules/tests
Round 11 review findings on PR #1218:

- find_opaque_struct_types's stable tier: old_by_stable_id/new_by_stable_id
  now map each StableEntityId to a list[RecordType] instead of overwriting
  duplicates. A snapshot can legitimately carry more than one declaration
  resolving to the same id (an ODR-duplicate pair, or an unreconciled
  header-AST TU-merge pair), and such a pair can disagree on is_opaque --
  a last-write-wins dict made the opacity verdict depend on iteration
  order. The decision is now per-id: when both sides have a bucket, the id
  is confirmed opaque only when every declaration in *both* lists is
  opaque; when exactly one side has a bucket, the existing asymmetric-
  absence check applies only once every declaration in that bucket is
  itself opaque.

Also closes the architecture-gate findings the round-10 fixes' own line
growth tripped (debt-no-growth on diff_filtering.py/post_processing.py,
new-file-size on opaque_types.py, new-test-size on two test files):

- abicheck/compare/opaque_struct_types.py (new): find_opaque_struct_types,
  split out of opaque_types.py (947 -> 716 lines).
- abicheck/compare/opaque_struct_downgrade.py (new): the whole struct-
  downgrade block (_OPAQUE_DOWNGRADEABLE, _struct_change_record_name,
  _resolve_struct_change_entity_id, _downgrade_opaque_struct_changes),
  split out of diff_filtering.py (1994 -> 1755 lines, under baseline)
  mirroring the existing base_class_diff.py/va_list_diff.py pattern.
  diff_filtering.py imports the public names back under their original
  private spellings, so internal call sites and external test imports are
  unaffected.
- post_processing.py: trimmed a comment to land exactly back at its
  1823-line no-growth baseline.
- tests/test_opaque_struct_identity_tiers.py (new): the ADR-063 Phase 10
  struct-downgrade/find_opaque_struct_types suites, split out of
  test_opaque_identity_tiers.py (1530 -> 623 lines split into 623 + 998),
  importing shared fixtures from the original file rather than duplicating
  them. Also trimmed one redundant test from test_cov95_diff_filtering.py
  (1219 -> 1196 lines) whose scenario is now covered, parametrized, in the
  new file.

New regression test (TestStableIdRequiresEveryDeclarationOpaque, both
duplicate orderings) confirmed to fail under the old last-write-wins dict.
