+++ abicheck/compare/opaque_types.py
@@ -66,7 +66,6 @@ if TYPE_CHECKING:
 __all__ = [
     "OpaqueTypeIndex",
     "find_by_value_types",
-    "find_opaque_struct_types",
     "find_opaque_types",
     "is_impl_source",
 ]
@@ -236,8 +235,8 @@ class OpaqueTypeIndex:
         case from a real single entity that is genuinely opaque on both
         sides or genuinely absent from the other side -- only the caller,
         which can resolve identity across both snapshots directly, can. See
-        :func:`find_opaque_struct_types` for the verification this performs
-        before passing a set here.
+        :func:`~abicheck.compare.opaque_struct_types.find_opaque_struct_types`
+        for the verification this performs before passing a set here.
 
         ``complete`` is left at its dataclass default (``True``) here on
         purpose, but that default is *not* a completeness proof the way
@@ -376,236 +375,6 @@ def find_opaque_types(snap: AbiSnapshot) -> OpaqueTypeIndex:
         local=frozenset(local),
         stable_by_local={k: frozenset(v) for k, v in stable_by_local.items()},
     )
-
-
-def find_opaque_struct_types(old: AbiSnapshot, new: AbiSnapshot) -> OpaqueTypeIndex:
-    """The DWARF-oriented, asymmetric-existence sibling of
-    :func:`find_opaque_types` -- ``diff_filtering._downgrade_opaque_struct_changes``'s
-    own opaqueness criterion, moved here (ADR-063 Phase 10) for the identical
-    reason :func:`find_opaque_types`/:func:`find_by_value_types` already sit
-    here rather than in ``diff_filtering.py``: a matching concern belongs at
-    its ADR-061 owner, and ``diff_filtering.py`` sits on a zero-slack
-    ``debt.yaml`` no-growth pin that this function's own logic (previously
-    inline there) would otherwise have exceeded.
-
-    A type is "opaque to consumers" under this function's criterion when:
-
-    * it is ``is_opaque`` on *both* sides, OR
-    * it is ``is_opaque`` on one side and entirely absent from the other
-      side's header-level type list (the DWARF-only-definition case: a
-      forward-declaration-only header paired with a snapshot that only saw
-      the type's full DWARF definition on the other side of the release)
-
-    and it is not embedded by value in any non-opaque record on either side.
-    Returns the empty index when nothing in *old*/*new* is opaque under this
-    criterion.
-
-    Deliberately does **not** reuse :func:`find_opaque_types`'s own
-    per-snapshot ``is_opaque``-or-``is_impl_source`` criterion or its
-    ``intersect()`` -- this function's asymmetric-existence rule (item 2
-    above) has no per-snapshot analogue to intersect two of, and mixing the
-    two criteria into one function would make either one harder to verify
-    against its own, independently-reviewed test suite. The two
-    opaque-suppression paths staying textually separate, each producing its
-    own :class:`OpaqueTypeIndex`, is the status quo this migration
-    preserves -- only *how* each already-computed name set answers "is this
-    ``Change`` about one of them" changed, not whether there are two paths.
-
-    ``opaque_types``/``embedded_types``/``truly_opaque`` stay plain
-    ``set[str]`` throughout -- both the asymmetric-existence check and the
-    by-value-embedding check are genuine rendered-text/bare-name-set
-    questions, not identity ones, mirroring why :func:`find_by_value_types`/
-    ``_type_is_by_value_referenced`` stay spelling-based too. Only the
-    *result* -- the final ``truly_opaque`` name set -- is re-expressed as a
-    two-tier index, via :meth:`OpaqueTypeIndex.build`, so a caller's own
-    ``Change`` membership test can consult a stable, cross-snapshot
-    ``EntityId`` first and the bare spelling second (never narrowing:
-    this index is not the product of a paired ``intersect()``, so it has
-    no completeness proof to license ``contains(..., strict=True)``).
-    """
-    old_opaque = {t.name for t in old.types if t.is_opaque}
-    new_opaque = {t.name for t in new.types if t.is_opaque}
-    old_type_names = {t.name for t in old.types}
-    new_type_names = {t.name for t in new.types}
-
-    opaque_types = (old_opaque & new_opaque) | (
-        (old_opaque - new_type_names) | (new_opaque - old_type_names)
-    )
-    if not opaque_types:
-        return OpaqueTypeIndex(stable=frozenset(), local=frozenset())
-
-    # Matched via :func:`_type_is_by_value_referenced` (the same
-    # qualification-robust helper :func:`find_by_value_types` already uses
-    # for function/variable signatures), not a bare
-    # ``f.type.rstrip(" *&") in opaque_types`` string comparison -- a field
-    # rendered as ``ns::Handle`` against a candidate whose own
-    # ``RecordType.name`` is bare ``"Handle"`` (or vice versa) would
-    # otherwise miss the embedding exclusion entirely (Codex review, PR
-    # #1218, round 3): "Handle" would stay in ``truly_opaque`` despite being
-    # embedded by value, and the stable tier -- which does not depend on
-    # this same spelling coincidence the way the pre-migration bare
-    # ``c.symbol`` comparison did -- would then downgrade a real, qualified
-    # layout-change finding for the embedded type.
-    #
-    # A round-5 review comment proposed narrowing a *bare* candidate's own
-    # match to the ``::``-excluding matcher (rejecting a match immediately
-    # preceded by ``::``), to stop an unrelated ``other::Handle`` field from
-    # wrongly excluding a bare ``"Handle"`` candidate as embedded. That
-    # narrowing was tried and reverted: it is textually indistinguishable
-    # from the round-3 case it would break -- a field rendered ``ns::Handle``
-    # against the *same* bare-named opaque candidate is not distinguishable,
-    # by string matching alone, from an unrelated ``other::Handle`` -- and
-    # the two failure directions are not symmetric. Under-matching here
-    # (round 5's direction) risks *silently suppressing* a genuine,
-    # consumer-visible layout break (the round-3 false negative this
-    # migration exists to close); over-matching (today's behavior) risks
-    # only an extra, spurious non-suppression -- a struct-size finding a
-    # human can still see and dismiss. AGENTS.md's "weaker evidence narrows
-    # conclusions... never upgrades to a clean compatibility claim" already
-    # states the correct default for this exact ambiguity: prefer not
-    # suppressing when the two directions cannot be told apart.
-    non_opaque_old = [t for t in old.types if not t.is_opaque]
-    non_opaque_new = [t for t in new.types if not t.is_opaque]
-    embedded_types: set[str] = set()
-    for records in (non_opaque_old, non_opaque_new):
-        for t in records:
-            for f in t.fields:
-                for tname in opaque_types - embedded_types:
-                    if _type_is_by_value_referenced(tname, f.type):
-                        embedded_types.add(tname)
-
-    truly_opaque = opaque_types - embedded_types
-    if not truly_opaque:
-        return OpaqueTypeIndex(stable=frozenset(), local=frozenset())
-
-    # Only a declaration that itself satisfies ``is_opaque`` may contribute
-    # to the local (bare-spelling) tier here -- the pre-migration
-    # `c.symbol in truly_opaque` fallback only ever compared against opaque
-    # declarations' own names, so a non-opaque namesake must not join
-    # `declarations` even for the always-safe spelling tier.
-    declarations: dict[str, list[RecordType]] = {}
-    for snap in (old, new):
-        for t in snap.types:
-            if t.name in truly_opaque and t.is_opaque:
-                declarations.setdefault(t.name, []).append(t)
-
-    # The *stable* tier needs a stricter, identity-paired check than "any
-    # opaque declaration under a qualifying spelling" -- two rounds of Codex
-    # review on PR #1218 found that grouping by bare name alone lets a
-    # stable id ride along on a coincidence rather than actual opacity:
-    #
-    # 1. An opaque ``ns::Handle`` and an unrelated, *visible* ``other::
-    #    Handle`` collide on the bare spelling "Handle" -- adding every
-    #    same-named declaration's id (not just the opaque one's) would
-    #    smuggle the visible declaration's own stable id into the index.
-    # 2. Even restricted to `is_opaque` declarations, ``old_opaque &
-    #    new_opaque`` only proves *some* declaration named "Handle" is
-    #    opaque on *each* side -- not that it is the *same* declaration.
-    #    If ``ns::Handle`` is opaque in old but goes visible in new, while
-    #    an unrelated ``other::Handle`` is opaque only in new, "Handle"
-    #    still lands in `truly_opaque` via the intersection, and
-    #    ``ns::Handle``'s own old-side stable id would wrongly enter the
-    #    index even though *that specific entity* is not opaque on both
-    #    sides at all.
-    #
-    # Resolving per stable EntityId across both snapshots (independent of
-    # bare-name grouping) closes both: an id may enter the stable tier only
-    # when the *same* entity is confirmed opaque under this function's own
-    # criterion --- opaque on both sides, or genuinely absent (by identity,
-    # not by name) from the other snapshot entirely.
-    old_by_stable_id: dict[StableEntityId, RecordType] = {}
-    for t in old.types:
-        resolved = stable_entity_id(t.entity_id)
-        if resolved is not None:
-            old_by_stable_id[resolved] = t
-    new_by_stable_id: dict[StableEntityId, RecordType] = {}
-    for t in new.types:
-        resolved = stable_entity_id(t.entity_id)
-        if resolved is not None:
-            new_by_stable_id[resolved] = t
-
-    # Every spelling (``name`` *and*, when present, ``qualified_name``) a
-    # declaration on each side is known by -- used only by the stable
-    # tier's own absence check just below, never by ``opaque_types``'s
-    # bare-name criterion above (that computation stays deliberately
-    # spelling-based/lossy per this function's own docstring; only the
-    # stable tier's identity-adjacent absence check needs the stricter
-    # multi-spelling view). A header-AST backend commonly stores the bare
-    # leaf in ``name`` and the real scoped spelling in ``qualified_name``
-    # (the same asymmetry ``_resolve_struct_change_entity_id``'s own
-    # round-9 fix closed for the sibling identity bridge) -- checking
-    # ``name`` alone here would misread a genuinely-present but
-    # identity-less counterpart spelled only via ``qualified_name`` as
-    # absent, letting an old opaque declaration's stable id enter the
-    # index and then get borrowed by that same counterpart's own
-    # ``qualified_name`` (Codex review, PR #1218, round 10).
-    old_all_spellings = old_type_names | {
-        t.qualified_name for t in old.types if t.qualified_name is not None
-    }
-    new_all_spellings = new_type_names | {
-        t.qualified_name for t in new.types if t.qualified_name is not None
-    }
-
-    # Check the stable counterpart BY ID first, before ever consulting bare
-    # names -- two producers can render the identical entity under two
-    # different spellings (this whole module's own premise: "the header
-    # backends key RecordType.name bare while DWARF bakes the namespace
-    # into name"), so a name-level absence check run first would misread a
-    # same-entity, differently-spelled counterpart as genuinely absent
-    # (Codex review, PR #1218, round 4): old's opaque bare "Handle" and
-    # new's now-visible "ns::Handle" can share one stable id despite never
-    # sharing a name at all. Only once the id itself is confirmed absent
-    # from ``other_by_id`` does bare-name presence decide anything -- and
-    # even then, a resolved id absent from ``other_by_id`` is not, by
-    # itself, proof the entity is absent from the other snapshot: a
-    # mixed-producer or pre-identity-baseline comparison can leave the
-    # other side's matching declaration present but carrying no resolvable
-    # ``entity_id`` at all (Codex review, PR #1218, round 3) -- missing
-    # identity *evidence* is not evidence of *absence*. The
-    # asymmetric-existence criterion this function implements is a
-    # NAME-level one in the first place (see ``opaque_types`` above, built
-    # from ``old_type_names``/``new_type_names``), so falling back to that
-    # same bare-name presence check only once identity itself is silent is
-    # not a new criterion, only a faithful re-check of the one already used
-    # to decide this name belongs in ``truly_opaque`` at all.
-    stable_ids: set[StableEntityId] = set()
-    for snap, other_all_spellings, other_by_id in (
-        (old, new_all_spellings, new_by_stable_id),
-        (new, old_all_spellings, old_by_stable_id),
-    ):
-        for t in snap.types:
-            if t.name not in truly_opaque or not t.is_opaque:
-                continue
-            resolved = stable_entity_id(t.entity_id)
-            if resolved is None:
-                continue
-            other_t = other_by_id.get(resolved)
-            if other_t is not None:
-                if other_t.is_opaque:
-                    # The same entity (by id, regardless of spelling),
-                    # confirmed opaque on both sides.
-                    stable_ids.add(resolved)
-                # Otherwise: the same entity is visible on the other side
-                # -- decline, regardless of what its name looks like there.
-                continue
-            known_spellings = {t.name}
-            if t.qualified_name is not None:
-                known_spellings.add(t.qualified_name)
-            if not (known_spellings & other_all_spellings):
-                # No entity anywhere on the other side resolves to this
-                # exact id, AND no declaration shares this entity's name OR
-                # qualified name either -- genuinely absent by both
-                # identity and every known spelling, the asymmetric-
-                # existence criterion.
-                stable_ids.add(resolved)
-            # Otherwise: a declaration under one of this entity's spellings
-            # exists on the other side, but this exact entity's own
-            # identity either did not resolve there or resolved to a
-            # different entity entirely -- decline the stable-tier match.
-            # The always-safe spelling tier still applies via
-            # ``declarations``/``local`` above.
-
-    return OpaqueTypeIndex.build(declarations, stable_ids=stable_ids)
 
 
 #: Matches whitespace or a leading cv-qualifier keyword, repeated -- the
